requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.1
coreschema==0.0.4
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
Django==1.11.5
django-cors-headers==2.1.0
django-filter==1.0.4
django-rest-swagger==2.1.2
djangorestframework==3.6.4
google-api-core==1.2.0
google-auth==1.5.0
google-cloud-core==0.28.1
google-cloud-translate==1.3.1
googleapis-common-protos==1.5.3
googletrans==2.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
itypes==1.1.0
Jinja2==2.9.6
kiwisolver==1.0.1
lxml==4.2.1
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.2.2
nltk==3.2.5
numpy==1.14.2
olefile==0.44
openapi-codec==1.3.2
pandas==0.23.0
parsel==1.4.0
Pillow==4.3.0
protobuf==3.5.2.post1
psycopg2==2.7.3.1
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
pyparsing==2.2.0
pypiwin32==223
python-crfsuite==0.9.5
python-dateutil==2.7.2
pytz==2018.4
pyvi==0.0.9.1
pywin32==223
queuelib==1.5.0
requests==2.18.4
rsa==3.4.2
scikit-learn==0.19.1
scipy==1.0.1
Scrapy==1.5.0
scrapy-djangoitem==1.1.1
service-identity==17.0.0
simplejson==3.11.1
six==1.11.0
sklearn-crfsuite==0.3.6
tabulate==0.8.2
tqdm==4.20.0
Twisted==17.9.0
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.22
w3lib==1.19.0
zope.interface==4.4.3
